provide constructor option to skip JSON conversion
by default, responses will be converted to JSON (from XML).  The
current converter doesn't handle collections uniformly.  If the collection
has 0 or 1 item, it becomes a JSON Object instead of an empty array or an
array of 1.

to be more backwards compatible, when you construct the VeracodeClient
you can simple provide a boolean flag that is used to bypass conversion.

by setting the default to `false`, current usages should remain unaffected